/test: move /test/e2e/lifecycle tests to /test/e2e_kubeadm
Move the orphaned (not run in any CI) tests from
/test/e2e/lifecycle to /test/e2e_kubeadm.

These tests will now run in all kinder jobs, some
of which are release-informing.

https://testgrid.k8s.io/sig-cluster-lifecycle-kubeadm

Replace the /test/e2e/lifecycle examples from the README.md
with /test/e2e/autoscaling ones.
