+++ test/e2e/e2e_test.go
@@ -56,8 +56,6 @@ import (
 	_ "k8s.io/kubernetes/test/e2e/invariants"
 	"k8s.io/kubernetes/test/e2e/invariants/logcheck"
 	_ "k8s.io/kubernetes/test/e2e/kubectl"
-	_ "k8s.io/kubernetes/test/e2e/lifecycle"
-	_ "k8s.io/kubernetes/test/e2e/lifecycle/bootstrap"
 	_ "k8s.io/kubernetes/test/e2e/network"
 	_ "k8s.io/kubernetes/test/e2e/node"
 	_ "k8s.io/kubernetes/test/e2e/scheduling"
