--- test/e2e/lifecycle/OWNERS
@@ -1,22 +1,0 @@
-# See the OWNERS docs at https://go.k8s.io/owners
-
-approvers:
-  - neolit123
-  - SataQiu
-  - pacoxu
-  - carlory
-  - HirazawaUi
-reviewers:
-  - neolit123
-  - SataQiu
-  - pacoxu
-  - carlory
-  - HirazawaUi
-labels:
-  - sig/cluster-lifecycle
-emeritus_approvers:
-  - fabriziopandini
-  - rosti
-  - ereslibre
-  - timothysc
-  - luxas
